+++ packages/samples/headless-react/src/pages/SearchPage.tsx
@@ -46,6 +46,8 @@ import {RelevanceInspector} from '../components/relevance-inspector/relevance-in
 import {RelevanceInspector as RelevanceInspectorFn} from '../components/relevance-inspector/relevance-inspector.fn';
 import {StandaloneSearchBox} from '../components/standalone-search-box/standalone-search-box.class';
 import {StandaloneSearchBox as StandaloneSearchBoxFn} from '../components/standalone-search-box/standalone-search-box.fn';
+import {RedirectionTrigger} from '../components/triggers/redirection-trigger.class';
+import {RedirectionTrigger as RedirectionTriggerFn} from '../components/triggers/redirection-trigger.fn';
 import {
   SearchEngine,
   buildSearchEngine,
@@ -97,6 +99,8 @@ import {
   buildRelevanceInspector,
   StandaloneSearchBox as HeadlessStandaloneSearchBox,
   buildStandaloneSearchBox,
+  RedirectionTrigger as HeadlessRedirectionTrigger,
+  buildRedirectionTrigger,
 } from '@coveo/headless';
 import {
   RecommendationEngine,
@@ -153,6 +157,7 @@ export class SearchPage extends Component {
   private readonly historyManager: HeadlessHistoryManager;
   private readonly relevanceInspector: HeadlessRelevanceInspector;
   private readonly standaloneSearchBox: HeadlessStandaloneSearchBox;
+  private readonly redirectionTrigger: HeadlessRedirectionTrigger;
 
   private unsubscribeUrlManager!: Unsubscribe;
 
@@ -274,6 +279,8 @@ export class SearchPage extends Component {
     this.standaloneSearchBox = buildStandaloneSearchBox(this.engine, {
       options: {redirectionUrl: 'https://mywebsite.com/search'},
     });
+
+    this.redirectionTrigger = buildRedirectionTrigger(this.engine);
   }
 
   componentDidMount() {
@@ -445,6 +452,12 @@ export class SearchPage extends Component {
           <Section title="standalone-search-box">
             <StandaloneSearchBox />
             <StandaloneSearchBoxFn controller={this.standaloneSearchBox} />
+          </Section>
+          <Section title="redirection-trigger">
+            <RedirectionTrigger></RedirectionTrigger>
+            <RedirectionTriggerFn
+              controller={this.redirectionTrigger}
+            ></RedirectionTriggerFn>
           </Section>
         </AppContext.Provider>
       </div>